+++ app/Http/Livewire/Ui/Form/Wrapper.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Http\Livewire\Ui\Form;
+
+use Livewire\Component;
+
+class Wrapper extends Component
+{
+    public function render()
+    {
+        return view('livewire.ui.form.wrapper');
+    }
+}