+++ app/Http/Requests/StorePostRequest.php
@@ -2,39 +2,57 @@
 
 namespace App\Http\Requests;
 
+use App\Traits\CanSanitizeInputs;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Str;
 use Illuminate\Validation\Rules\File;
 
-class StorePostRequest extends FormRequest
-{
-    public function authorize(): bool
-    {
-        return Gate::has('admin');
-    }
+/**
+ * @property string $tags
+ * @property string $title
+ * @property string $slug
+ */
+class StorePostRequest extends FormRequest {
+	use CanSanitizeInputs;
+
+	public function authorize(): bool
+	{
+		return Gate::has('admin');
+	}
 
 	public function prepareForValidation(): void
 	{
 		$this->merge([
+			'categories' => collect($this->categories)
+				->map(fn ($category) => $this->sanitize($category))
+				->filter()
+				->values()
+				->map(static fn ($category) => (int) $category)
+				->toArray(),
 			'slug' => Str::slug($this->slug ?? $this->title ?? ''),
+			'tags' => $this->sanitize($this->tags ?? ''),
 		]);
 	}
 
-    public function rules(): array
-    {
-        return [
+	public function rules(): array
+	{
+		return [
 			'cover_image' => File::image()->max(5 * 1024),
-            'title' => [
+			'title' => [
 				'required',
 				'max:180',
 			],
+			'search_title' => 'max:180',
+			'search_description' => 'max:250',
 			'slug' => [
 				'required',
 				'unique:posts',
 				'max:180',
 			],
 			'body' => 'required',
-        ];
-    }
+			'categories' => 'array',
+			'tags' => 'string',
+		];
+	}
 }