+++ app/Models/Category.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Casts\Attribute;
+use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
@@ -16,7 +17,7 @@ use Illuminate\Support\Str;
  * @property string $slug
  * @property \Illuminate\Support\Carbon|null $created_at
  * @property \Illuminate\Support\Carbon|null $updated_at
- * @property-read \Illuminate\Database\Eloquent\Collection|\App\Models\Post[] $posts
+ * @property-read Collection|\App\Models\Post[] $posts
  * @property-read int|null $posts_count
  * @method static \Database\Factories\CategoryFactory factory(...$parameters)
  * @method static \Illuminate\Database\Eloquent\Builder|Category newModelQuery()
@@ -31,6 +32,16 @@ use Illuminate\Support\Str;
 class Category extends Model
 {
     use HasFactory;
+
+	protected $fillable = [
+		'title',
+		'slug',
+	];
+
+	public static function fromIds(array $IDs): Collection
+	{
+		return static::whereIn('id', collect($IDs)->unique())->get();
+	}
 
 	public function posts(): MorphToMany
 	{