+++ app/Models/Post.php
@@ -2,17 +2,20 @@
 
 namespace App\Models;
 
-use App\Contracts\ImageableContract;
+use App\Contracts\CategorizeableContract;
+use App\Contracts\SearchDisplayableContract;
 use App\Enums\PerPage;
-use App\Traits\Imageable;
+use App\Traits\Categorizeable;
+use App\Traits\SearchDisplayable;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use Illuminate\Support\Carbon;
 use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Http\Request;
 use Illuminate\Pagination\AbstractPaginator;
+use Illuminate\Support\Collection;
+use Spatie\Sitemap\Contracts\Sitemapable;
+use Spatie\Sitemap\Tags\Url;
 
 /**
  * App\Models\Post
@@ -24,7 +27,7 @@ use Illuminate\Pagination\AbstractPaginator;
  * @property string $excerpt
  * @property Image|null $image
  * @property Category[] $categories
- * @property Tag[] $tags
+ * @property \Illuminate\Database\Eloquent\Collection|Tag[] $tags
  * @property Carbon $created_at
  * @property Carbon $updated_at
  * @property-read int|null $categories_count
@@ -43,10 +46,17 @@ use Illuminate\Pagination\AbstractPaginator;
  * @mixin \Eloquent
  * @property-read \Illuminate\Database\Eloquent\Collection|\App\Models\Image[] $images
  * @property-read int|null $images_count
+ * @property string $page_title
+ * @property string $meta_description
+ * @property \App\Models\SearchMeta $searchMeta
  */
-class Post extends ImageableModel
+class Post extends ImageableModel implements SearchDisplayableContract, CategorizeableContract, Sitemapable
 {
-    use HasFactory;
+	// TODO: Implement Taggable stuff too
+	// TODO: Replace ImageableModel stuff with contract and trait ONLY
+    use Categorizeable,
+		HasFactory,
+		SearchDisplayable;
 
 	/**
 	 * @var string[]
@@ -57,11 +67,6 @@ class Post extends ImageableModel
 		'slug',
 		'title',
 	];
-
-	public function categories(): MorphToMany
-	{
-		return $this->morphToMany(Category::class, 'categorizeable');
-	}
 
 	public function excerpt(): Attribute
 	{
@@ -94,5 +99,64 @@ class Post extends ImageableModel
 				optional($request)->per_page
 			))
 			->withQueryString();
+	}
+
+	public function pageTitle(): Attribute
+	{
+		return Attribute::make(
+			get: fn () => $this->searchMeta?->search_title ?? $this->title,
+
+			// TODO: how can this be more efficient?
+			set: function (string $title) {
+				$meta = $this->searchMeta()->firstOrCreate();
+				$meta->update(['search_title' => $title]);
+			},
+		);
+	}
+
+	public function metaDescription(): Attribute
+	{
+		return Attribute::make(
+			get: fn () => $this->searchMeta?->search_description ?? str($this->body)
+				->words(30, '')
+				->replaceMatches("/(\r\n|\r|\n)+/", " ")
+				->whenEndsWith(['.', '?', '!', '...'],
+					static fn ($string) => $string->append(''),
+					static fn ($string) => $string->append('...'),
+				)->toString(),
+
+			// TODO: how can this be more efficient?
+			set: function (string $description) {
+				$meta = $this->searchMeta()->firstOrCreate();
+				$meta->update(['search_description' => $description]);
+			},
+		);
+	}
+
+	/**
+	 * @param \Illuminate\Support\Collection<\App\Models\Category|int> $categories
+	 * @return self
+	 */
+	public function syncCategories(Collection $categories): self
+	{
+		$this->categories()->sync($categories->map(static fn (Category $category) => $category->id)->all());
+
+		return $this;
+	}
+
+	/**
+	 * @param \Illuminate\Support\Collection<\App\Models\Tag> $tags
+	 * @return self
+	 */
+	public function syncTags(Collection $tags): self
+	{
+		$this->tags()->sync($tags->map(static fn (Tag $tag) => $tag->id)->all());
+
+		return $this;
+	}
+
+	public function toSitemapTag(): Url|string|array
+	{
+		return route('blog.show', ['post' => $this]);
 	}
 }